Add pooling logic for weapon

+++ Assets/_Game/Scripts/Data/UserData.cs
@@ -19,8 +19,8 @@ public class UserData
     }
 
     // Player stats.
-    [SerializeField] private int _level = 1;
-    public int Level { get => _level; set => _level = value; }
+    [SerializeField] private byte _level = 1;
+    public byte Level { get => _level; set => _level = value; }
 
     [SerializeField] private int _experience = 0;
     public int Experience { get => _experience; set => _experience = value; }
@@ -39,24 +39,24 @@ public class UserData
     public int HighScore { get => _highScore; set => _highScore = value; }
 
     // Unlocked items.
-    [SerializeField] private List<int> _unlockedWeapons = new List<int>();
-    public List<int> UnlockedWeapons { get => _unlockedWeapons; set => _unlockedWeapons = value; }
+    [SerializeField] private List<byte> _unlockedWeapons = new List<byte>();
+    public List<byte> UnlockedWeapons { get => _unlockedWeapons; set => _unlockedWeapons = value; }
 
-    [SerializeField] private List<int> _unlockedPants = new List<int>();
-    public List<int> UnlockedPants { get => _unlockedPants; set => _unlockedPants = value; }
+    [SerializeField] private List<byte> _unlockedPants = new List<byte>();
+    public List<byte> UnlockedPants { get => _unlockedPants; set => _unlockedPants = value; }
 
-    [SerializeField] private List<int> _unlockedHairs = new List<int>();
-    public List<int> UnlockedHairs { get => _unlockedHairs; set => _unlockedHairs = value; }
+    [SerializeField] private List<byte> _unlockedHairs = new List<byte>();
+    public List<byte> UnlockedHairs { get => _unlockedHairs; set => _unlockedHairs = value; }
 
     // Equipped items.
-    [SerializeField] private int _equippedWeapon = 0;
-    public int EquippedWeapon { get => _equippedWeapon; set => _equippedWeapon = value; }
+    [SerializeField] private byte _equippedWeapon = 0;
+    public byte EquippedWeapon { get => _equippedWeapon; set => _equippedWeapon = value; }
 
-    [SerializeField] private int _equippedPant = 0;
-    public int EquippedPant { get => _equippedPant; set => _equippedPant = value; }
+    [SerializeField] private byte _equippedPant = 0;
+    public byte EquippedPant { get => _equippedPant; set => _equippedPant = value; }
 
-    [SerializeField] private int _equippedHair = 0;
-    public int EquippedHair { get => _equippedHair; set => _equippedHair = value; }
+    [SerializeField] private byte _equippedHair = 0;
+    public byte EquippedHair { get => _equippedHair; set => _equippedHair = value; }
 
     public UserData()
     {
